Api data classes: change id from Int to Long due to occuring exception

+++ app/src/main/java/com/example/plantidentificationapp/classes/IdentifiedPlant.kt
@@ -2,7 +2,7 @@ package com.example.plantidentificationapp.classes
 
 import com.example.plantidentificationapp.api.model.PlantDetails
 
-class IdentifiedPlant(var id: Int,
+class IdentifiedPlant(var id: Long,
                       var plant_details: PlantDetails,
                       var plant_name: String,
                       var probability: Double)
